Plan 9 from Bell Labs 1992-03-09

+++ port/portdat.h
@@ -46,7 +46,7 @@ struct List
 struct Ref
 {
 	Lock;
-	int	ref;
+	long	ref;
 };
 
 struct Rendez
